+++ omnipath/_core/requests/_annotations.py
@@ -1,4 +1,5 @@
-from typing import Any, Dict, Mapping
+from typing import Any, Dict, Union, Mapping, Iterable, Optional
+import logging
 
 import pandas as pd
 
@@ -6,6 +7,8 @@ from omnipath._core.query import QueryType
 from omnipath._core.utils._docs import d
 from omnipath._core.requests._request import OmnipathRequestABC
 from omnipath.constants._pkg_constants import Key, final
+
+_MAX_N_PROTS = 600
 
 
 @final
@@ -34,6 +37,9 @@ class Annotations(OmnipathRequestABC):
     @classmethod
     def get(
         cls,
+        proteins: Optional[Union[str, Iterable[str]]] = None,
+        resources: Optional[Union[str, Iterable[str]]] = None,
+        force_full_download: bool = False,
         **kwargs,
     ) -> pd.DataFrame:
         """
@@ -49,6 +55,15 @@ class Annotations(OmnipathRequestABC):
 
             In order to download annotations for proteins complexes, write **'COMPLEX:'** before the gene symbols of
             the genes integrating the complex.
+
+            If `None`, fetch annotations for all available genes or proteins.
+        resources
+            Load the annotations only from these databases. See :meth:`resources` for available options.
+            If `None`, use all available resources.
+        force_full_download
+            Force the download of the entire annotations dataset. The full size of the data is ~1GB.
+            We recommend to retrieve the annotations for a set of proteins or only from a few resources,
+            depending on your interest.
         kwargs
             Additional query parameters.
 
@@ -62,37 +77,41 @@ class Annotations(OmnipathRequestABC):
         There might be also a few miRNAs annotated. A vast majority of protein complex annotations are inferred
         from the annotations of the members: if all members carry the same annotation the complex inherits.
         """
-        # fmt: off
-        if not any(
-            kw in kwargs
-            for kw in ("proteins", "resources", "databases")
-        ):
-
+        if proteins is None and resources is None and not force_full_download:
             raise ValueError(
-                "Downloading the entire annotations database is not allowed "
-                "due to its huge size (>1GB). Please query a set of proteins "
-                "or a few resources, depending on your interest."
+                "Please specify `force_full_download=True` in order to download the full dataset."
             )
-        # fmt: on
+        res_info = (
+            "all resources"
+            if resources is None
+            else f"the following resources: `{[resources] if isinstance(resources, str) else sorted(set(resources))}`"
+        )
+        inst = cls()
 
-        if "proteins" in kwargs:
-
-            proteins = kwargs["proteins"]
+        if proteins is not None:
             if isinstance(proteins, str):
                 proteins = (proteins,)
-            proteins = tuple(set(proteins))
+            proteins = sorted(set(proteins))
 
-            if len(proteins) > 600:
-                # fmt: off
-                raise ValueError(
-                    "Cannot download annotations for "
-                    "more than `600` proteins yet."
-                )
-                # fmt: on
+            logging.info(
+                f"Downloading annotations for `{len(proteins)}` in `{_MAX_N_PROTS}` chunks from {res_info}"
+            )
 
-            kwargs["proteins"] = proteins
+            return pd.concat(
+                [
+                    inst._get(
+                        proteins=proteins[i * _MAX_N_PROTS : (i + 1) * _MAX_N_PROTS],
+                        resources=resources,
+                        **kwargs,
+                    )
+                    for i in range((len(proteins) // _MAX_N_PROTS) + 1)
+                    if len(proteins[i * _MAX_N_PROTS : (i + 1) * _MAX_N_PROTS])
+                ]
+            )
 
-        return cls()._get(**kwargs)
+        logging.info(f"Downloading annotations for all proteins from {res_info}")
+
+        return inst._get(proteins=None, resources=resources, **kwargs)
 
     def _resource_filter(self, data: Mapping[str, Any], **_) -> bool:
         return True